Yet another file location changed.
Sorry, I can't help much because I'm a bit of a hack myself. But this script is a touch... brittle.

And it requires starting from scratch every time. It would be lovely if for each library it could do a check to see if the library exists before doing the download/compile, and maybe save the cleanup until the end.

At least that way if a library location does change, we don't have to start from the beginning every time. Ok, maybe I'll dust off my bash and try to figure out a fix in a future PR.